# webapp/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/manager/seats" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SeatDTO"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          }
        },
        "summary" : "Get All Manager Seats",
        "tags" : [ "Seat Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "ids",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "204" : {
            "description" : "Seat deleted successfully"
          },
          "400" : {
            "description" : "Bad Request: Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Seat with specified ID not found for the current manager"
          }
        },
        "summary" : "Delete Manager Seat",
        "tags" : [ "Seat Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SeatRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SeatDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Event location not found"
          },
          "409" : {
            "description" : "Conflict: Seat with this row and number already exists in this event location"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Seat",
        "tags" : [ "Seat Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/seats/{id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SeatRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SeatDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Seat with specified ID not found for the current manager"
          },
          "409" : {
            "description" : "Conflict: Seat with this row and number already exists in this event location"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update Manager Seat",
        "tags" : [ "Seat Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SeatDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Seat with specified ID not found for the current manager"
          }
        },
        "summary" : "Get Manager Seat By Id",
        "tags" : [ "Seat Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SeatDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seatNumber" : {
            "type" : "string"
          },
          "seatRow" : {
            "type" : "string"
          },
          "locationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "xCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "yCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entrance" : {
            "type" : "string"
          }
        }
      },
      "SeatRequestDTO" : {
        "type" : "object",
        "required" : [ "seatNumber", "eventLocationId", "xCoordinate", "yCoordinate", "seatRow" ],
        "properties" : {
          "seatNumber" : {
            "type" : "string"
          },
          "eventLocationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "xCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "yCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "seatRow" : {
            "type" : "string"
          },
          "entrance" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
...
  "info" : {
    "title" : "Seat Reservation API",
    "version" : "0.0.7",
    "description" : "The official REST API for Seat Reservation.",
    "contact" : {
      "name" : "FelixHertweck GitHub ",
      "url" : "https://github.com/FelixHertweck"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}